+++ plugins/ingestion-kafka/src/main/java/org/opensearch/plugin/kafka/KafkaOffset.java
@@ -39,11 +39,6 @@ public class KafkaOffset implements IngestionShardPointer {
     }
 
     @Override
-    public String asString() {
-        return String.valueOf(offset);
-    }
-
-    @Override
     public Field asPointField(String fieldName) {
         return new LongPoint(fieldName, offset);
     }
@@ -63,7 +58,7 @@ public class KafkaOffset implements IngestionShardPointer {
 
     @Override
     public String toString() {
-        return "KafkaOffset{" + "offset=" + offset + '}';
+        return String.valueOf(offset);
     }
 
     @Override