+++ server/src/main/java/org/opensearch/action/admin/indices/streamingingestion/resume/TransportResumeIngestionAction.java
@@ -112,36 +112,65 @@ public class TransportResumeIngestionAction extends TransportClusterManagerNodeA
             return;
         }
 
+        ActionListener<UpdateIngestionStateResponse> stateUpdateListener = new ActionListener<>() {
+
+            @Override
+            public void onResponse(UpdateIngestionStateResponse updateIngestionStateResponse) {
+                boolean shardsAcked = updateIngestionStateResponse.isAcknowledged() && updateIngestionStateResponse.getFailedShards() == 0;
+                ResumeIngestionResponse response = new ResumeIngestionResponse(
+                    true,
+                    shardsAcked,
+                    updateIngestionStateResponse.getShardFailureList(),
+                    updateIngestionStateResponse.getErrorMessage()
+                );
+                listener.onResponse(response);
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                logger.debug("Error resuming ingestion", e);
+                listener.onFailure(e);
+            }
+        };
+
         String[] indices = Arrays.stream(concreteIndices).map(Index::getName).toArray(String[]::new);
+        if (request.getResetSettings() != null && request.getResetSettings().length > 0) {
+            // reset consumer and resume ingestion
+            UpdateIngestionStateRequest shardPointerUpdateRequest = getShardPointerUpdateRequest(indices, request);
+            UpdateIngestionStateRequest resumeIngestionRequest = getIngestionResumeRequest(indices, request);
+            ingestionStateService.resetShardPointerAndResumeIngestion(
+                "resume-ingestion",
+                concreteIndices,
+                shardPointerUpdateRequest,
+                resumeIngestionRequest,
+                stateUpdateListener
+            );
+        } else {
+            // resume ingestion
+            UpdateIngestionStateRequest updateIngestionStateRequest = getIngestionResumeRequest(indices, request);
+            ingestionStateService.updateIngestionPollerState(
+                "resume-ingestion",
+                concreteIndices,
+                updateIngestionStateRequest,
+                stateUpdateListener
+            );
+        }
+    }
+
+    private UpdateIngestionStateRequest getShardPointerUpdateRequest(String[] indices, ResumeIngestionRequest request) {
+        int[] shards = Arrays.stream(request.getResetSettings()).mapToInt(ResumeIngestionRequest.ResetSettings::getShard).toArray();
+        UpdateIngestionStateRequest updateIngestionStateRequest = new UpdateIngestionStateRequest(indices, shards);
+        updateIngestionStateRequest.timeout(request.clusterManagerNodeTimeout());
+        updateIngestionStateRequest.setResetSettings(request.getResetSettings());
+
+        return updateIngestionStateRequest;
+    }
+
+    private UpdateIngestionStateRequest getIngestionResumeRequest(String[] indices, ResumeIngestionRequest request) {
         UpdateIngestionStateRequest updateIngestionStateRequest = new UpdateIngestionStateRequest(indices, new int[0]);
         updateIngestionStateRequest.timeout(request.clusterManagerNodeTimeout());
         updateIngestionStateRequest.setIngestionPaused(false);
 
-        ingestionStateService.updateIngestionPollerState(
-            "resume-ingestion",
-            concreteIndices,
-            updateIngestionStateRequest,
-            new ActionListener<>() {
-
-                @Override
-                public void onResponse(UpdateIngestionStateResponse updateIngestionStateResponse) {
-                    boolean shardsAcked = updateIngestionStateResponse.isAcknowledged()
-                        && updateIngestionStateResponse.getFailedShards() == 0;
-                    ResumeIngestionResponse response = new ResumeIngestionResponse(
-                        true,
-                        shardsAcked,
-                        updateIngestionStateResponse.getShardFailureList(),
-                        updateIngestionStateResponse.getErrorMessage()
-                    );
-                    listener.onResponse(response);
-                }
-
-                @Override
-                public void onFailure(Exception e) {
-                    logger.debug("Error resuming ingestion", e);
-                    listener.onFailure(e);
-                }
-            }
-        );
+        return updateIngestionStateRequest;
     }
 }